% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basemap}
\alias{basemap}
\title{A base map of the world.
Country names are renamed to match the data from covariants}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 244 rows and 2 columns.
}
\usage{
basemap
}
\description{
A base map of the world.
Country names are renamed to match the data from covariants
}
\keyword{datasets}
