% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_list}
\alias{country_list}
\title{A list of all countries available in the data.}
\format{
An object of class \code{character} of length 146.
}
\usage{
country_list
}
\description{
A list of all countries available in the data.
}
\keyword{datasets}
