% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dates_list}
\alias{dates_list}
\title{A list of all available dates in the global data}
\format{
An object of class \code{Date} of length 52.
}
\usage{
dates_list
}
\description{
A list of all available dates in the global data
}
\keyword{datasets}
