% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main.R
\name{heatmap_theme}
\alias{heatmap_theme}
\title{Heatmap Theme
this is the base theme to apply to all of the heatmaps}
\usage{
heatmap_theme(lineage)
}
\arguments{
\item{lineage}{name of lineage used to identify fill in legend}
}
\description{
Heatmap Theme
this is the base theme to apply to all of the heatmaps
}
