% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main.R
\name{map_dominant_variant}
\alias{map_dominant_variant}
\title{Map Dominan Variant
map based on the most prevalent variant for a given week}
\usage{
map_dominant_variant(weekz)
}
\arguments{
\item{weekz}{week ending}
}
\value{
a plotly plot
}
\description{
Map Dominan Variant
map based on the most prevalent variant for a given week
}
