% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main.R
\name{map_variant}
\alias{map_variant}
\title{Map Variant
map based on a selected variant for a given week}
\usage{
map_variant(.data, pangoz, weekz, quintile = TRUE)
}
\arguments{
\item{pangoz}{variant lineage}

\item{weekz}{week ending}

\item{quantile}{whether or not to split data into quintiles}
}
\value{
a plotly plot
}
\description{
Map Variant
map based on a selected variant for a given week
}
