% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omicron_global_proportions}
\alias{omicron_global_proportions}
\title{A list of percentage of total covid cases are each individual variant
combined across all countries}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 918 rows and 4 columns.
}
\usage{
omicron_global_proportions
}
\description{
A list of percentage of total covid cases are each individual variant
combined across all countries
}
\keyword{datasets}
