% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omicron_proportions}
\alias{omicron_proportions}
\title{A list of percentage of total covid cases are each individual variant
stratified by week and country}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13977 rows and 5 columns.
}
\usage{
omicron_proportions
}
\description{
A list of percentage of total covid cases are each individual variant
stratified by week and country
}
\keyword{datasets}
