% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main.R
\name{plot_heatmap_by_country}
\alias{plot_heatmap_by_country}
\title{Plot Heatmap by Country
plots a heatmap by country and date for a specific variant}
\usage{
plot_heatmap_by_country(lineage = "BA.1")
}
\arguments{
\item{lineage}{the omicron lineage to filter by}
}
\value{
a plotly plot
}
\description{
Plot Heatmap by Country
plots a heatmap by country and date for a specific variant
}
