% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main.R
\name{plot_heatmap_by_lineage}
\alias{plot_heatmap_by_lineage}
\title{Plot Heatmap By Lineage
plots a heatmap of proportion of each variant globally to give a sense of dominant variant over time.}
\usage{
plot_heatmap_by_lineage()
}
\description{
Plot Heatmap By Lineage
plots a heatmap of proportion of each variant globally to give a sense of dominant variant over time.
}
