% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main.R
\name{plot_lineage}
\alias{plot_lineage}
\title{Plot Lineage
plots a heatmap by country and date for a specific variant}
\usage{
plot_lineage(lineage = "BA.1")
}
\arguments{
\item{lineage}{the omicron lineage to filter by}
}
\value{
a plotly plot
}
\description{
Plot Lineage
plots a heatmap by country and date for a specific variant
}
