% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main.R
\name{plot_proportions}
\alias{plot_proportions}
\title{Plot Proportions}
\usage{
plot_proportions(countriez = "USA")
}
\arguments{
\item{countriez}{The country to filter by}
}
\value{
a plotly plot
}
\description{
Plots the proportions of each variant of all variants for a given country over time.
}
