% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main.R
\name{prep_maps}
\alias{prep_maps}
\title{Prepare Maps
Merge dataframe with world geometries on country column}
\usage{
prep_maps(.data)
}
\arguments{
\item{.data}{data to merge with country map}
}
\description{
Prepare Maps
Merge dataframe with world geometries on country column
}
