% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variant_list}
\alias{variant_list}
\title{A list of all available variants}
\format{
An object of class \code{factor} of length 18.
}
\usage{
variant_list
}
\description{
A list of all available variants
}
\keyword{datasets}
